% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runClusterExperiment.R
\name{runClusterExperiment}
\alias{runClusterExperiment}
\title{Run the shiny app for clusterExperiment package}
\usage{
runClusterExperiment()
}
\description{
This function runs the shiny app that corresponds to the clusterExperiment
package.
}
\details{
Typing \code{runClusterExperiment()} at the command prompt (after
  loading the \code{clusterExpShiny} library) will cause a browser to open
  with the shiny app.
}

